Fix indentation and alignment of lines 94 and 95
While the previous commit replaced some horizontal tabs with sets of
spaces, it did not replace all of them. Further, the else statement on
line 94 remained misaligned with the preceding if statement on line 92.
As a result, the code failed to run. This fixes both the inconsistent
indentation style and the misaligned conditional statements.